Fix broken ModbusSerialClient port/baudrate arg handling
**Why?** My previous PR https://github.com/Salamek/epsolar-tracer/commit/78bb985a5bb5
fixed the handling of other keyword args like timeout, but broke the `port`/`baudrate`
args because those args were explicitly extracted from `kwargs` and passed
into `ModbusSerialClient`. Thus when either `port` or `baudrate` were present
in `kwargs`, they were duplicated in the arguments passed to
`ModbusSerialClient` which results in an exception.

To avoid that, this commit stops passing the explicit `port`/`baudrate`
args into `ModbusSerialClient` and opts to preserve the existing behavior of
having defaults for those values by writing the defaults into `kwargs`
iff they are not present there.

One could argue that it is a little unconventional to modify `kwargs`,
however it seems to be the most concise way to keep the code simple
while preserving the existing external interface of
`EPsolarTracerClient`. Ref: https://stackoverflow.com/a/45883925